Replace Database.getCanvas with Database.getRow
I removed the Database.getCanvas method I added in commit 1aa8643 and
replaced it with a Database.getRow method. My initial plan was to
implement the functionality of retrieving the whole canvas at once
directly inside the database module. This turned out to be a bit chaotic
(especially because I tried returning the result as a Buffer, which
didn't work correctly as Node handles string-to-buffer conversions a bit
weird), so I removed the method and replaced it with a new method to
just retrieve one row at a time. Concatenating all rows together should
then be handled by action_canvas(), which I'll implement soon.