Add check for usage on Windows
I added code into src/index.ts to check if the script is executed using
a Node version compiled on/for the Microsoft Windows platform. As this
project is designed with UNIX-based systems in mind (Linux, macOS, etc.)
flawless support for non-UNIX platforms, especially Win32, can not be
guaranteed. When being started on Windows, the script will output a
log message explaining the situation and warning the user about
potential issues when using Huechan Place on non-UNIX platforms.