Add draw cooldown
Added implementation for the draw cooldown. When the user tries to draw
a tile, the code checks if there is a cooldown already active for the
user. If there is one active, the action is aborted and a failure
response is sent. If there's not, the action completes and the cooldown
is being started. The current implementation also support distinguishing
between the regular draw cooldown and the special mode cooldown.
(ConfigurationKeys.AlbaniaCooldown)