Implement consistent naming of Special mode
Although Special mode is referred to with this name in the README and in
code added in the last few weeks, the name previously used internally
was 'Albania mode'. I started this project with the specific use case to
use it for personal purposes, so I modelled the Special mode around my
specific implementation. Later, this changed and I decided to leave the
implementation/interpretation of what the special tiles should represent
to be done client/frontend-side. I now replaced all remnants of
'Albania mode' with the new labels.