Implement action_info
Implemented action_info in src/info.ts. Now by requesting the "info"
action, the client will receive the size of the canvas, the time
remaining until the cooldown ends, a timestamp of when the Special mode
starts (ConfigurationKeys.SpecialModeStartup) and a timestamp of when
the server will terminate operation. (ConfigurationKeys.Shutdown)